Reduced advantage of moving diagonally

+++ core/src/com/nick/ant/towerdefense/renderables/entities/Entity.java
@@ -9,26 +9,26 @@ import com.nick.ant.towerdefense.renderables.rooms.Room;
  * Created by Nick on 08/09/2014.
  */
 public abstract class Entity extends Renderable {
-    protected int x = 0;
-    protected int y = 0;
-    protected int hSpeed = 0;
-    protected int vSpeed = 0;
+    protected float x = 0;
+    protected float y = 0;
+    protected float hSpeed = 0;
+    protected float vSpeed = 0;
     protected Room room;
     protected CollisionManager collisionManager;
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
@@ -46,8 +46,8 @@ public abstract class Entity extends Renderable {
             return;
         }
 
-        int newX = x + hSpeed;
-        int newY = y + vSpeed;
+        float newX = x + hSpeed;
+        float newY = y + vSpeed;
 
         if (collisionManager == null)   {
             x = newX;
